main.rs implemented, waiting... for Raspi 5 to get fixed to test

+++ src/config.rs
@@ -22,7 +22,7 @@ pub struct LeetcodeConfig {
 
 #[derive(Deserialize)]
 pub struct UnlockConfig {
-    pub duration_minutes: i32,
+    pub duration_minutes: u64,
 }
 
 #[derive(Deserialize)]
